Split auto-type emission into a focused mixin.
